requirements.txt
Django==5.1
Pillow>=10.0.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# For MySQL in production (optional):
# mysqlclient>=2.2.0
